+++ app/demo/page.tsx
@@ -1,0 +1,15 @@
+import type { Metadata } from "next"
+import { ViewerIntro } from "@/components/viewer-intro"
+
+export const metadata: Metadata = {
+  title: "Visualizador 3D | LAMBDA 3D",
+  description: "Explora nuestros modelos 3D interactivos para educación médica y planificación quirúrgica",
+}
+
+export default function VisualizadorPage() {
+  return (
+    <main className="min-h-screen">
+      <ViewerIntro />
+    </main>
+  )
+}
