+++ app/globals.css
@@ -14,75 +14,93 @@ html, body, :root { height: 100%; }
 /* (resto igual) */
 @layer base {
   :root {
-    --background: 0 0% 100%;
-    --foreground: 0 0% 3.9%;
+    /* ===== PALETA INSTITUCIONAL =====
+       Azul:   #33809d  → hsl(196 51% 41%)
+       Negro:  #1d1d1b  → hsl(60 4% 11%)
+       Blanco: #ffffff  → hsl(0 0% 100%)
+       Gris:   #575756  → hsl(60 1% 34%)
+    */
+    --background: 0 0% 100%;              /* blanco */
+    --foreground: 60 4% 11%;              /* negro */
+
+    --primary: 196 51% 41%;               /* azul marca */
+    --primary-foreground: 0 0% 100%;      /* blanco */
+
+    /* secundarios/accent derivados de la marca */
+    --secondary: 196 32% 92%;
+    --secondary-foreground: 60 4% 11%;
+    --accent: 196 32% 92%;
+    --accent-foreground: 60 4% 11%;
+    --muted: 196 32% 95%;
+    --muted-foreground: 60 1% 34%;
+
+    /* gris institucional para borders/inputs */
+    --border: 60 1% 84%;                  /* gris claro (derivado) */
+    --input: 60 1% 84%;
+    --ring: 196 51% 41%;                  /* foco con azul marca */
+
+    /* tarjetas/popovers */
     --card: 0 0% 100%;
-    --card-foreground: 0 0% 3.9%;
+    --card-foreground: 60 4% 11%;
     --popover: 0 0% 100%;
-    --popover-foreground: 0 0% 3.9%;
-    --primary: 0 0% 9%;
-    --primary-foreground: 0 0% 98%;
-    --secondary: 0 0% 96.1%;
-    --secondary-foreground: 0 0% 9%;
-    --muted: 0 0% 96.1%;
-    --muted-foreground: 0 0% 45.1%;
-    --accent: 0 0% 96.1%;
-    --accent-foreground: 0 0% 9%;
-    --destructive: 0 84.2% 60.2%;
-    --destructive-foreground: 0 0% 98%;
-    --border: 0 0% 89.8%;
-    --input: 0 0% 89.8%;
-    --ring: 0 0% 3.9%;
-    --chart-1: 12 76% 61%;
-    --chart-2: 173 58% 39%;
-    --chart-3: 197 37% 24%;
-    --chart-4: 43 74% 66%;
-    --chart-5: 27 87% 67%;
+    --popover-foreground: 60 4% 11%;
+
+    /* estados */
+    --destructive: 0 84% 60%;
+    --destructive-foreground: 0 0% 100%;
+
+    /* radios/shadows que ya uses */
     --radius: 0.5rem;
+
+    /* (opcional) sidebar si lo usás */
     --sidebar-background: 0 0% 98%;
-    --sidebar-foreground: 240 5.3% 26.1%;
-    --sidebar-primary: 240 5.9% 10%;
-    --sidebar-primary-foreground: 0 0% 98%;
-    --sidebar-accent: 240 4.8% 95.9%;
-    --sidebar-accent-foreground: 240 5.9% 10%;
-    --sidebar-border: 220 13% 91%;
-    --sidebar-ring: 217.2 91.2% 59.8%;
-  }
-  .dark {
-    --background: 0 0% 3.9%;
-    --foreground: 0 0% 98%;
-    --card: 0 0% 3.9%;
-    --card-foreground: 0 0% 98%;
-    --popover: 0 0% 3.9%;
-    --popover-foreground: 0 0% 98%;
-    --primary: 0 0% 98%;
-    --primary-foreground: 0 0% 9%;
-    --secondary: 0 0% 14.9%;
-    --secondary-foreground: 0 0% 98%;
-    --muted: 0 0% 14.9%;
-    --muted-foreground: 0 0% 63.9%;
-    --accent: 0 0% 14.9%;
-    --accent-foreground: 0 0% 98%;
-    --destructive: 0 62.8% 30.6%;
-    --destructive-foreground: 0 0% 98%;
-    --border: 0 0% 14.9%;
-    --input: 0 0% 14.9%;
-    --ring: 0 0% 83.1%;
-    --chart-1: 220 70% 50%;
-    --chart-2: 160 60% 45%;
-    --chart-3: 30 80% 55%;
-    --chart-4: 280 65% 60%;
-    --chart-5: 340 75% 55%;
-    --sidebar-background: 240 5.9% 10%;
-    --sidebar-foreground: 240 4.8% 95.9%;
-    --sidebar-primary: 224.3 76.3% 48%;
+    --sidebar-foreground: 60 4% 11%;
+    --sidebar-primary: 196 51% 41%;
     --sidebar-primary-foreground: 0 0% 100%;
-    --sidebar-accent: 240 3.7% 15.9%;
-    --sidebar-accent-foreground: 240 4.8% 95.9%;
-    --sidebar-border: 240 3.7% 15.9%;
-    --sidebar-ring: 217.2 91.2% 59.8%;
+    --sidebar-accent: 196 32% 92%;
+    --sidebar-accent-foreground: 60 4% 11%;
+    --sidebar-border: 60 1% 84%;
+    --sidebar-ring: 196 51% 41%;
+  }
+
+  .dark {
+    /* versión dark manteniendo la identidad */
+    --background: 224 71% 4%;
+    --foreground: 0 0% 96%;
+
+    --card: 224 71% 6%;
+    --card-foreground: 0 0% 96%;
+    --popover: 224 71% 6%;
+    --popover-foreground: 0 0% 96%;
+
+    --primary: 196 52% 64%;               /* azul levantado para contraste */
+    --primary-foreground: 222 47% 11%;
+
+    --secondary: 215 28% 17%;
+    --secondary-foreground: 0 0% 96%;
+    --muted: 215 28% 17%;
+    --muted-foreground: 0 0% 80%;
+    --accent: 215 28% 17%;
+    --accent-foreground: 0 0% 96%;
+
+    --destructive: 0 63% 31%;
+    --destructive-foreground: 0 0% 96%;
+
+    --border: 215 28% 17%;
+    --input: 215 28% 17%;
+    --ring: 196 52% 64%;
+
+    --sidebar-background: 240 6% 10%;
+    --sidebar-foreground: 0 0% 96%;
+    --sidebar-primary: 196 52% 64%;
+    --sidebar-primary-foreground: 222 47% 11%;
+    --sidebar-accent: 240 4% 16%;
+    --sidebar-accent-foreground: 0 0% 96%;
+    --sidebar-border: 240 4% 16%;
+    --sidebar-ring: 196 52% 64%;
   }
 }
+
 
 .lil-gui.root { right: 14px !important; top: 120px !important; }
