+++ app/visualizador/admin/globalsadmin.css
