+++ app/visualizador/admin/layout.tsx
