+++ app/visualizador/admin/page.tsx
@@ -5,17 +5,17 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/com
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Toaster } from "@/components/ui/toaster"
 import { Users, FolderOpen, Shield } from "lucide-react"
-import { UsersTable, type User, type Project } from "@/components/admin/UsersTable"
-import { ProjectsTable } from "@/components/admin/ProjectsTable"
-import { PermissionsEditor } from "@/components/admin/PermissionsEditor"
-import { ImportExportACL } from "@/components/admin/ImportExportACL"
-import { FirebaseUserManager } from "@/components/admin/FirebaseUserManager"
-import { FirebaseProjectManager } from "@/components/admin/FirebaseProjectManager"
+import { UsersTable, type User, type Project } from "@/components/visualizador/admin/UsersTable"
+import { ProjectsTable } from "@/components/visualizador/admin/ProjectsTable"
+import { PermissionsEditor } from "@/components/visualizador/admin/PermissionsEditor"
+import { ImportExportACL } from "@/components/visualizador/admin/ImportExportACL"
+import { FirebaseUserManager } from "@/components/visualizador/admin/FirebaseUserManager"
+import { FirebaseProjectManager } from "@/components/visualizador/admin/FirebaseProjectManager"
 import type { UserRole } from "@/lib/firebase"
 import { getUsers, type FirebaseUser } from "@/lib/firebase"
 import { getProjects } from "@/lib/firebase"
-import {RefreshWeb} from "@/components/admin/refreshWeb"
-import { HomeButton } from "@/components/UserLinks"
+import {RefreshWeb} from "@/components/visualizador/admin/refreshWeb"
+import { HomeButton } from "@/components/visualizador/UserLinks"
 import { ThemeToggle } from "@/components/ThemeToggle"
 import { getACL, type ACL } from "@/lib/acl"
