+++ app/visualizador/layout.tsx
@@ -1,0 +1,24 @@
+import type { ReactNode } from "react";
+import type { Metadata } from "next";
+import { GeistSans } from "geist/font/sans";
+import { GeistMono } from "geist/font/mono";
+import { ThemeProvider } from "@/components/theme-provider";
+// import "@/globals.css";
+
+export const metadata: Metadata = {
+  title: "3D visualizer",
+  description: "3D visualizer",
+  generator: "v0.dev",
+};
+
+export default function RootLayout({ children }: { children: ReactNode }) {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body className={`min-h-dvh font-sans ${GeistSans.variable} ${GeistMono.variable}`}>
+        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+          {children}
+        </ThemeProvider>
+      </body>
+    </html>
+  );
+}
