+++ components/visualizador/ControlsPanel.tsx
