+++ components/visualizador/UserLinks.tsx
@@ -7,7 +7,7 @@ import { ShieldAlertIcon,Home } from "lucide-react";
 export function GoToAdminLink() {
   return (
     
-      <Link href="/admin" className="inline-flex w-full items-center gap-2">
+      <Link href="/visualizador/admin" className="inline-flex w-full items-center gap-2">
             <ShieldAlertIcon className="mr-2 h-4 w-4 text-blue-500" />
             <span className="text-blue-500">Admin</span>
       </Link>
@@ -19,7 +19,7 @@ export function GoToAdminLink() {
 export function HomeButton() {
   return (
     <Button asChild variant="outline" size="sm">
-      <Link href="/">
+      <Link href="/visualizador">
         <Home className="h-[1.2rem] w-[1.2rem]" />
         <span className="sr-only">Ir al inicio</span>
       </Link>
