+++ components/visualizador/admin/AdminButton.tsx
