+++ components/visualizador/admin/PermissionsEditor.tsx
